+++ src/Main.java
@@ -1,4 +1,4 @@
-import model.Service;
+import model.service.Service;
 import presenter.Presenter;
 import view.ConsoleUI;
 import view.View;
